Refactored UIInputSystemManager
