--- Where_Is_My_Wife/Assets/_Main/Scripts/Input/DestroyIncorrectUIEventModule.cs
@@ -1,23 +1,0 @@
-using UnityEngine;
-using UnityEngine.EventSystems;
-using UnityEngine.InputSystem.UI;
-
-namespace WhereIsMyWife.Porting
-{
-    public class DestroyIncorrectUIEventModule : MonoBehaviour
-    {
-        [SerializeField] private StandaloneInputModule _standaloneInputModule;
-        [SerializeField] private InputSystemUIInputModule _inputSystemUiInputModule;
-    
-        private void Awake()
-        {
-        #if XBOX
-            Destroy(_inputSystemUiInputModule);
-        #endif
-                
-        #if !XBOX
-            Destroy(_standaloneInputModule);
-        #endif  
-        }
-    }
-}
